+++ lib/dart_jsonwebtoken.dart
@@ -4,3 +4,4 @@ export 'src/jwt.dart';
 export 'src/errors.dart';
 export 'src/algorithms.dart';
 export 'src/keys.dart';
+export 'src/parser.dart';